refactor(export): locate raw identity with analyzer
